clean up Avatar component
