+++ apps/client/src/pages/user/[id].tsx
@@ -2,20 +2,18 @@ import { createProxySSGHelpers } from "@trpc/react-query/ssg";
 import { GetServerSidePropsContext, InferGetServerSidePropsType } from "next";
 import Link from "next/dist/client/link";
 import Head from "next/dist/shared/lib/head";
-import Image from "next/image";
 import { useRouter } from "next/router";
 import { useEffect } from "react";
 
 import { useSession } from "../../client/auth/useSession";
 import { trpc } from "../../client/trpc";
+import Avatar from "../../home/Avatar";
 import MetaTags from "../../home/MetaTags";
 import { getNavbarLayout } from "../../home/NavbarLayout/NavbarLayout";
-import ProfilePicture from "../../home/ProfilePicture";
 import SpaceCard from "../../home/SpaceCard";
 import { prisma } from "../../server/prisma";
 import { appRouter } from "../../server/router/_app";
 import Spinner from "../../ui/Spinner";
-import { isFromCDN } from "../../utils/isFromCDN";
 import { hexDisplayToNumber, numberToHexDisplay } from "../../utils/numberToHexDisplay";
 
 export const getServerSideProps = async ({ res, query }: GetServerSidePropsContext) => {
@@ -128,31 +126,21 @@ export default function User({ id }: InferGetServerSidePropsType<typeof getServe
         <div className="max-w-content mx-auto">
           <div className="h-48 w-full bg-neutral-200 md:h-64 lg:rounded-xl">
             <div className="relative h-full w-full object-cover">
-              {profile?.metadata?.animation_url &&
-                (isFromCDN(profile.metadata.animation_url) ? (
-                  <Image
-                    src={profile.metadata.animation_url}
-                    priority
-                    fill
-                    sizes="80vw"
-                    alt=""
-                    className="h-full w-full object-cover lg:rounded-xl"
-                  />
-                ) : (
-                  <img
-                    src={profile.metadata.animation_url}
-                    alt=""
-                    className="h-full w-full object-cover lg:rounded-xl"
-                    crossOrigin="anonymous"
-                  />
-                ))}
+              {profile?.metadata?.animation_url && (
+                <img
+                  src={profile.metadata.animation_url}
+                  alt=""
+                  className="h-full w-full object-cover lg:rounded-xl"
+                  crossOrigin="anonymous"
+                />
+              )}
             </div>
           </div>
 
           <section className="flex justify-center px-4 pb-6 md:px-0">
             <div className="flex w-full flex-col items-center space-y-2">
               <div className="z-10 -mt-16 flex w-32 rounded-full ring-4 ring-white">
-                <ProfilePicture
+                <Avatar
                   src={profile?.metadata?.image}
                   circle
                   uniqueKey={profile?.handle?.full ?? id}
